Fix platform bindings

+++ platform/android/MapLibreAndroid/src/main/java/org/maplibre/android/maps/NativeMapOptions.java
@@ -57,7 +57,7 @@ public class NativeMapOptions {
     return actionJournalLogFileCount;
   }
 
-  public long actionJournalRenderingReportInterval() {
+  public int actionJournalRenderingReportInterval() {
     return actionJournalRenderingReportInterval;
   }
 }
